requirements.txt
autopep8==1.6.0
click==8.0.3
colorama==0.4.4
numpy==1.22.1
opencv-python==4.5.5.62
Pillow==9.0.0
pycodestyle==2.8.0
pyzbar==0.1.8
qrcode==7.3.1
shellingham==1.4.0
toml==0.10.2
typer==0.4.0
